+++ tests/lib/js/export/inheritance_collision_lib.dart
@@ -1,12 +1,10 @@
-// Copyright (c) 2022, the Dart project authors.  Please see the AUTHORS file
+// Copyright (c) 2025, the Dart project authors.  Please see the AUTHORS file
 // for details. All rights reserved. Use of this source code is governed by a
 // BSD-style license that can be found in the LICENSE file.
 
 // Tests collisions of `@JSExport` members using inheritance.
 
 import 'dart:js_interop';
-
-import 'package:js/js_util.dart';
 
 // Overridden members do not count as an export name collision.
 @JSExport()
@@ -99,32 +97,32 @@ mixin MixinCollision2 {
 
 @JSExport()
 class InheritanceRenameOneCollision extends SuperclassCollision
-//    ^
-// [web] The following class members collide with the same export 'field': MixinCollision.fieldMixin, SuperclassCollision.fieldSuper.
-// [web] The following class members collide with the same export 'finalField': MixinCollision.finalFieldMixin, SuperclassCollision.finalFieldSuper.
-// [web] The following class members collide with the same export 'getSet': MixinCollision.getSetMixin, MixinCollision.getSetMixin, SuperclassCollision.getSetSuper, SuperclassCollision.getSetSuper.
-// [web] The following class members collide with the same export 'method': MixinCollision.methodMixin, SuperclassCollision.methodSuper.
+        // [error column 7]
+        // [web] The following class members collide with the same export 'field': MixinCollision.fieldMixin, SuperclassCollision.fieldSuper.
+        // [web] The following class members collide with the same export 'finalField': MixinCollision.finalFieldMixin, SuperclassCollision.finalFieldSuper.
+        // [web] The following class members collide with the same export 'getSet': MixinCollision.getSetMixin, MixinCollision.getSetMixin, SuperclassCollision.getSetSuper, SuperclassCollision.getSetSuper.
+        // [web] The following class members collide with the same export 'method': MixinCollision.methodMixin, SuperclassCollision.methodSuper.
         with
         MixinCollision {}
 
 @JSExport()
 class InheritanceRenameTwoCollisions extends SuperclassCollision
-//    ^
-// [web] The following class members collide with the same export 'field': MixinCollision.fieldMixin, MixinCollision2.fieldMixin2, SuperclassCollision.fieldSuper.
-// [web] The following class members collide with the same export 'finalField': MixinCollision.finalFieldMixin, MixinCollision2.finalFieldMixin2, SuperclassCollision.finalFieldSuper.
-// [web] The following class members collide with the same export 'getSet': MixinCollision.getSetMixin, MixinCollision.getSetMixin, MixinCollision2.getSetMixin2, MixinCollision2.getSetMixin2, SuperclassCollision.getSetSuper, SuperclassCollision.getSetSuper.
-// [web] The following class members collide with the same export 'method': MixinCollision.methodMixin, MixinCollision2.methodMixin2, SuperclassCollision.methodSuper.
+        // [error column 7]
+        // [web] The following class members collide with the same export 'field': MixinCollision.fieldMixin, MixinCollision2.fieldMixin2, SuperclassCollision.fieldSuper.
+        // [web] The following class members collide with the same export 'finalField': MixinCollision.finalFieldMixin, MixinCollision2.finalFieldMixin2, SuperclassCollision.finalFieldSuper.
+        // [web] The following class members collide with the same export 'getSet': MixinCollision.getSetMixin, MixinCollision.getSetMixin, MixinCollision2.getSetMixin2, MixinCollision2.getSetMixin2, SuperclassCollision.getSetSuper, SuperclassCollision.getSetSuper.
+        // [web] The following class members collide with the same export 'method': MixinCollision.methodMixin, MixinCollision2.methodMixin2, SuperclassCollision.methodSuper.
         with
         MixinCollision,
         MixinCollision2 {}
 
 @JSExport()
 class InheritanceRenameThreeCollisions extends SuperclassCollision
-//    ^
-// [web] The following class members collide with the same export 'field': InheritanceRenameThreeCollisions.fieldDerived, MixinCollision.fieldMixin, MixinCollision2.fieldMixin2, SuperclassCollision.fieldSuper.
-// [web] The following class members collide with the same export 'finalField': InheritanceRenameThreeCollisions.finalFieldDerived, MixinCollision.finalFieldMixin, MixinCollision2.finalFieldMixin2, SuperclassCollision.finalFieldSuper.
-// [web] The following class members collide with the same export 'getSet': InheritanceRenameThreeCollisions.getSetDerived, InheritanceRenameThreeCollisions.getSetDerived, MixinCollision.getSetMixin, MixinCollision.getSetMixin, MixinCollision2.getSetMixin2, MixinCollision2.getSetMixin2, SuperclassCollision.getSetSuper, SuperclassCollision.getSetSuper.
-// [web] The following class members collide with the same export 'method': InheritanceRenameThreeCollisions.methodDerived, MixinCollision.methodMixin, MixinCollision2.methodMixin2, SuperclassCollision.methodSuper.
+        // [error column 7]
+        // [web] The following class members collide with the same export 'field': InheritanceRenameThreeCollisions.fieldDerived, MixinCollision.fieldMixin, MixinCollision2.fieldMixin2, SuperclassCollision.fieldSuper.
+        // [web] The following class members collide with the same export 'finalField': InheritanceRenameThreeCollisions.finalFieldDerived, MixinCollision.finalFieldMixin, MixinCollision2.finalFieldMixin2, SuperclassCollision.finalFieldSuper.
+        // [web] The following class members collide with the same export 'getSet': InheritanceRenameThreeCollisions.getSetDerived, InheritanceRenameThreeCollisions.getSetDerived, MixinCollision.getSetMixin, MixinCollision.getSetMixin, MixinCollision2.getSetMixin2, MixinCollision2.getSetMixin2, SuperclassCollision.getSetSuper, SuperclassCollision.getSetSuper.
+        // [web] The following class members collide with the same export 'method': InheritanceRenameThreeCollisions.methodDerived, MixinCollision.methodMixin, MixinCollision2.methodMixin2, SuperclassCollision.methodSuper.
         with
         MixinCollision,
         MixinCollision2 {
@@ -186,33 +184,4 @@ class GetSet {
 class GetSetInheritanceNoCollision extends Fields {
   int get setter => throw '';
   set getter(int val) => throw '';
-}
-
-void main() {
-  createDartExport(InheritanceOneOverrideNoCollision());
-  createDartExport(InheritanceTwoOverridesNoCollision());
-  createDartExport(InheritanceThreeOverridesNoCollision());
-
-  createDartExport(InheritanceRenameOneCollision());
-  createDartExport(InheritanceRenameTwoCollisions());
-  createDartExport(InheritanceRenameThreeCollisions());
-
-  createDartExport(InheritanceNoSuperclassMembers());
-
-  createDartExport(PartialOverrideFieldNoCollision());
-  createDartExport(GetSetInheritanceNoCollision());
-
-  // Same method with different name and type.
-  createJSInteropWrapper(InheritanceOneOverrideNoCollision());
-  createJSInteropWrapper(InheritanceTwoOverridesNoCollision());
-  createJSInteropWrapper(InheritanceThreeOverridesNoCollision());
-
-  createJSInteropWrapper(InheritanceRenameOneCollision());
-  createJSInteropWrapper(InheritanceRenameTwoCollisions());
-  createJSInteropWrapper(InheritanceRenameThreeCollisions());
-
-  createJSInteropWrapper(InheritanceNoSuperclassMembers());
-
-  createJSInteropWrapper(PartialOverrideFieldNoCollision());
-  createJSInteropWrapper(GetSetInheritanceNoCollision());
 }